\name{batch}
\Rdversion{1.1}
\alias{batch}
\title{
  Stores information on the batch (e.g., chemistry plate) used to
  process samples.
}
\description{
  Parameters for estimating copy number are locus- and batch-specific.
}
\usage{
batch(object)
}
\arguments{
  \item{object}{An object of class \code{CrlmmSetList} or \code{CopyNumberSet}.
  }
}
\value{
  Typically, a character string or factor.
}
\author{
  Rob Scharpf
}
\keyword{manip}

