\name{ffdf-class}
\Rdversion{1.1}
\docType{class}
\alias{ffdf-class}
\alias{annotatedDataFrameFrom,ff_matrix-method}
\alias{list-class}
\title{Class "ffdf"}
\description{
	object of class \code{ffdf}
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Slots}{
  \describe{
    \item{\code{.S3Class}:}{Object of class \code{"character"} ~~ }
  }
}
\section{Extends}{
Class \code{"\linkS4class{oldClass}"}, directly.
Class \code{"\linkS4class{list_or_ffdf}"}, directly.
Class \code{"\linkS4class{ff_or_matrix}"}, directly.
}
\section{Methods}{
No methods defined with class "ffdf" in the signature.
}
\seealso{
	\code{link{ffdf}}
}
\examples{
showClass("ffdf")
}
\keyword{classes}
