\name{totalCopyNumber-methods}
\docType{methods}
\alias{totalCopyNumber}
\alias{totalCopyNumber-methods}
\alias{totalCopyNumber,CNSet-method}
\title{Method for computing total copy number from either ff}
\description{
	calculates the sum of the allele-specific CN estimates
}
\section{Methods}{
\describe{

\item{\code{signature(object = "CNSet")}}{
  an object of class \code{CNSet} or \code{CNSetLM}.			     			     			     
}
}}
\keyword{methods}
