\name{validCdfNames}
\alias{validCdfNames}
\title{
	Valid R package for annotating crlmm or illumina platforms
}
\description{
	Returns a list of the platforms that are currently supported.
}
\details{
	
}
\value{
	Vector of currently supported platforms.
}
\references{
	}
\author{
	R. Scharpf
}


\examples{
	validCdfNames()
}
\keyword{manip}

